refactor(backend): import path-access helpers directly, remove re-exports
All consumers now import checkAccess/isSubPath from @/helpers/path-access.js
directly instead of going through intermediate re-exports.
