+++ apps/backend/src/agent/tools/file/search-files.ts
@@ -14,8 +14,9 @@ import type {
   ToolExecuteResult,
   ToolExecutionContext,
 } from '@/agent-core/tool/index.js';
+import {AccessCheckResult, checkAccess} from '@/helpers/path-access.js';
 
-import {AccessCheckResult, checkAccess, isBinaryFile} from './helpers.js';
+import {isBinaryFile} from './helpers.js';
 
 const MAX_MATCHES = 100;
 const MAX_CONCURRENCY = 10;
